Fix empty channel double rendering bug
Double clicking on the same server would cause the first channel on the list to not display.
When the expectation was to always display the first available channel. A simple check to
the server ID in the URL before it is pushed prevents throwing the user in the "no channels"
component by not having a :channelId in the address.